switch to inspect callback
